+++ parsedmarc/kafkaclient.py
@@ -1,10 +1,13 @@
-#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+
+import logging
+import json
 
 from kafka import KafkaProducer
 from kafka.errors import NoBrokersAvailable, UnknownTopicOrPartitionError
-import json
 
+
+logger = logging.getLogger("parsedmarc")
 
 class KafkaError(RuntimeError):
     """Raised when a Kafka error occurs"""
@@ -13,12 +16,12 @@ class KafkaError(RuntimeError):
 class KafkaClient(object):
     def __init__(self, kafka_hosts):
         try:
-            def serializer(v): lambda v: json.dumps(v).encode('utf-8')
             self.producer = KafkaProducer(
-                          value_serializer=serializer,
+                          value_serializer=lambda v: json.dumps(v).encode(
+                              'utf-8'),
                           bootstrap_servers=kafka_hosts)
         except NoBrokersAvailable:
-            raise KafkaError("No Kafka brokers availabe")
+            raise KafkaError("No Kafka brokers available")
 
     def save_aggregate_reports_to_kafka(self, aggregate_reports,
                                         aggregate_topic):
@@ -27,7 +30,8 @@ class KafkaClient(object):
 
         Args:
             aggregate_reports (list):  A list of aggregate report dictionaries
-            to save to kafka
+            to save to Kafka
+            aggregate_topic (str): The name of the Kafka topic
 
         """
         if type(aggregate_reports) == dict:
@@ -37,10 +41,19 @@ class KafkaClient(object):
             return
 
         try:
+            logger.debug("Saving aggregate reports to Kafka")
             self.producer.send(aggregate_topic, aggregate_reports)
         except UnknownTopicOrPartitionError:
-                raise KafkaError("Unknown topic or partition on broker")
-        self.producer.flush()
+                raise KafkaError(
+                    "Kafka error: Unknown topic or partition on broker")
+        except Exception as e:
+            raise KafkaError(
+                "Kafka error: {0}".format(e.__str__()))
+        try:
+            self.producer.flush()
+        except Exception as e:
+            raise KafkaError(
+                "Kafka error: {0}".format(e.__str__()))
 
     def save_forensic_reports_to_kafka(self, forensic_reports, forensic_topic):
             """
@@ -48,7 +61,8 @@ class KafkaClient(object):
 
             Args:
                 forensic_reports (list):  A list of forensic report dicts
-                to save to kafka
+                to save to Kafka
+                forensic_topic (str): The name of the Kafka topic
 
             """
             if type(forensic_reports) == dict:
@@ -58,7 +72,16 @@ class KafkaClient(object):
                 return
 
             try:
+                logger.debug("Saving forensic reports to Kafka")
                 self.producer.send(forensic_topic, forensic_reports)
             except UnknownTopicOrPartitionError:
-                raise KafkaError("Unknown topic or partition on broker")
-            self.producer.flush()
+                raise KafkaError(
+                    "Kafka error: Unknown topic or partition on broker")
+            except Exception as e:
+                raise KafkaError(
+                    "Kafka error: {0}".format(e.__str__()))
+            try:
+                self.producer.flush()
+            except Exception as e:
+                raise KafkaError(
+                    "Kafka error: {0}".format(e.__str__()))